auth-service/pyproject.toml
[tool.poetry]
name = "auth-service"
version = "0.1.0"
description = ""
authors = ["man-droid23 <alifilmannafian6b@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
asyncpg = "^0.27.0"
fastapi = "^0.87.0"
sqlmodel = "^0.0.8"
uvicorn = "^0.20.0"
alembic = "^1.8.1"
pytest = "^7.2.0"
pytest-asyncio = "^0.20.2"
httpx = "^0.23.0"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
